styling mostly complete
i need to stop burning time on working through all of the quirks!

there seems to be a bug where the first row fills the whole scrollbar
viewport, instead of the space it's supposed to occupy.

the ui is close-enough.. now to finish it off, and start using
"openai_dive"